[UPDATE] folders structure and refactor buttons styles
